% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_B.R
\name{pred_B}
\alias{pred_B}
\title{Population dynamics model to estimate gross fish biomass production}
\usage{
pred_B(
  GR = NULL,
  N0 = NULL,
  M_ref = NULL,
  L_ref = NULL,
  Lm = NULL,
  K = NULL,
  Linf = NULL,
  t0 = NULL,
  a = NULL,
  b = NULL,
  L0 = NULL,
  t.steps = NULL,
  plot = FALSE
)
}
\arguments{
\item{GR}{Fish growth rate during the juvenile phase (mm/day)}

\item{N0}{Initial number of fish in the cohort}

\item{M_ref}{Reference mortality rate (yr-1)}

\item{L_ref}{Reference size at which M = M_ref (in mm)}

\item{Lm}{Size at maturity (mm)}

\item{K}{von Bertalanffy growth rate (yr-1)}

\item{Linf}{von Bertalanffy asymptotic size (mm)}

\item{t0}{von Bertalanffy theoretical age at length zero (unitless)}

\item{a}{Intercept of length-weight relationship in the form W = aL^b}

\item{b}{Slope of length-weight relationship in the form W = aL^b}

\item{L0}{Initial size / size at hatching (mm)}

\item{t.steps}{Number of time steps to simulate population dynamics (in days)}

\item{plot}{Logical ('TRUE' or 'FALSE')}
}
\value{
Simulated biomass trajectories
}
\description{
Converts population demographic rates into cohort accumulated biomass
}
\examples{
pred_B(GR = 0.5, M_ref = 1.6, L_ref = 131.5, Lm = 131.5, K = 0.33, Linf = 219,
t0 = -1.1, t.steps = 365*10, L0 = 20, N0 = 0.4, a = 0.0064, b = 2.6,
plot = TRUE, K_sup = 10)
}
