% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Derivative/perturbation-based sensitivity analysis}
\usage{
sensitivity(
  par_to_vary,
  input_range = c(0, 100),
  baseline = list(mean, low, up),
  GR.mu = NULL,
  GR.sd = NULL,
  N0.mu = NULL,
  N0.sd = NULL,
  M_ref.mu = NULL,
  M_ref.sd = NULL,
  L_ref.mu = NULL,
  L_ref.sd = NULL,
  Lm.mu = NULL,
  Lm.sd = NULL,
  K.mu = NULL,
  K.sd = NULL,
  Linf.mu = NULL,
  Linf.sd = NULL,
  t0.mu = NULL,
  t0.sd = NULL,
  lwa.mu = NULL,
  lwa.sd = NULL,
  lwb.mu = NULL,
  lwb.sd = NULL,
  L0.mu = NULL,
  L0.sd = NULL,
  N_years = NULL,
  N_sim = 1000,
  N_ite = 1000
)
}
\arguments{
\item{par_to_vary}{Target input parameter (options are 'GR', 'K', 'Linf', 'Lm')}

\item{baseline}{Baseline input parameter value. Format is '=c(mean, lower, upper)'}

\item{GR.mu}{Mean fish growth rate during the juvenile phase (mm/day)}

\item{GR.sd}{SD of fish growth rate during the juvenile phase (mm/day)}

\item{N0.mu}{Mean initial number of fish in the cohort}

\item{N0.sd}{SD of initial number of fish in the cohort}

\item{M_ref.mu}{Mean reference mortality rate (yr-1)}

\item{M_ref.sd}{SD of reference mortality rate (yr-1)}

\item{L_ref.mu}{Mean reference size at which M = M_ref (in mm)}

\item{L_ref.sd}{SD of reference size at which M = M_ref (in mm)}

\item{Lm.mu}{Mean size at maturity (mm)}

\item{Lm.sd}{SD of size at maturity (mm)}

\item{K.mu}{Mean von Bertalanffy growth rate (yr-1)}

\item{K.sd}{SD of von Bertalanffy growth rate (yr-1)}

\item{Linf.mu}{Mean von Bertalanffy asymptotic size (mm)}

\item{Linf.sd}{SD of von Bertalanffy asymptotic size (mm)}

\item{t0.mu}{Mean von Bertalanffy theoretical age at length zero (unitless)}

\item{t0.sd}{SD of von Bertalanffy theoretical age at length zero (unitless)}

\item{lwa.mu}{Mean intercept of length-weight relationship in the form W = aL^b}

\item{lwa.sd}{SD of intercept of length-weight relationship in the form W = aL^b}

\item{lwb.mu}{Mean slope of length-weight relationship in the form W = aL^b}

\item{lwb.sd}{SD of slope of length-weight relationship in the form W = aL^b}

\item{L0.mu}{Mean initial size / size at hatching (mm)}

\item{L0.sd}{SD of initial size / size at hatching (mm)}

\item{N_years}{Number of time steps to project the cohort (in years)}

\item{N_sim}{Number of repeated Monte Carlo procedures}

\item{N_ite}{Number of iterations for each Monte Carlo procedure}

\item{range}{Range of input parameter to examine sensitivity of outputs. Format is '=c(0.1,0.6)'}

\item{plot}{Logical ('TRUE' or 'FALSE')}
}
\value{
Simulated gross biomass production with upper and lower bounds, % scaled values, and sensitivity index
}
\description{
Estimates model sensitivity to changes in input parameters
}
\examples{
sensitivity(par_to_vary = "GR", input_range = c(0.4,0.6), 
baseline = list(mean = 0.5, low = 0.45, up = 0.55),
GR.mu = 0.45, GR.sd = 0.03, N0.mu = 1, N0.sd = 0.01,
M_ref.mu = 1.6, M_ref.sd = 0.05, L_ref.mu = 120, L_ref.sd = 5,
Lm.mu = 120, Lm.sd = 5, K.mu = 0.33, K.sd = 0.1, Linf.mu = 200,
Linf.sd = 30, t0.mu = 0, t0.sd = 0, lwa.mu = 0.005,
lwa.sd = 0.000000001, lwb.mu = 3.25, lwb.sd = 0.1,
L0.mu = 20, L0.sd = 0, N_sim = 5, N_ite = 100)
}
